fix: add suspense to search
