+++ core/app/[locale]/(default)/(faceted)/search/page.tsx
@@ -1,11 +1,11 @@
-import { getFormatter, getTranslations } from 'next-intl/server';
+import { useTranslations } from 'next-intl';
+import { getTranslations } from 'next-intl/server';
+import { Suspense } from 'react';
 
-import { CompareProducts } from '@/vibes/soul/components/compare-products';
+import { CompareProductsSkeleton } from '@/vibes/soul/components/compare-products';
 import { SearchForm } from '~/components/search-form';
-import { facetsTransformer } from '~/data-transformers/facets-transformer';
-import { pricesTransformer } from '~/data-transformers/prices-transformer';
 
-import { fetchFacetedSearch } from '../fetch-faceted-search';
+import { ProductList } from './_components/product-list';
 
 export async function generateMetadata() {
   const t = await getTranslations('Search');
@@ -19,9 +19,8 @@ interface Props {
   searchParams: Record<string, string | string[] | undefined>;
 }
 
-export default async function Search({ searchParams }: Props) {
-  const t = await getTranslations('Search');
-  const format = await getFormatter();
+export default function Search({ searchParams }: Props) {
+  const t = useTranslations('Search');
 
   const searchTerm = typeof searchParams.term === 'string' ? searchParams.term : undefined;
 
@@ -34,78 +33,10 @@ export default async function Search({ searchParams }: Props) {
     );
   }
 
-  const search = await fetchFacetedSearch({ ...searchParams });
-
-  const productsCollection = search.products;
-  const products = productsCollection.items.map((product) => ({
-    id: product.entityId.toString(),
-    name: product.name,
-    href: product.path,
-    image: product.defaultImage
-      ? { src: product.defaultImage.url, altText: product.defaultImage.altText }
-      : undefined,
-    price: pricesTransformer(product.prices, format),
-    subtitle: product.brand?.name ?? undefined,
-  }));
-
-  if (products.length === 0) {
-    return (
-      <div>
-        <SearchForm initialTerm={searchTerm} />
-      </div>
-    );
-  }
-
-  const totalProducts = productsCollection.collectionInfo?.totalItems ?? 0;
-
-  const { hasNextPage, hasPreviousPage, endCursor, startCursor } = productsCollection.pageInfo;
-
-  const facets = search.facets.items.filter((facet) => facet.__typename !== 'CategorySearchFilter');
-  const filters = await facetsTransformer(facets);
-
-  const selectedSort = searchParams.sort ?? 'featured';
-
   return (
-    <CompareProducts
-      breadcrumbs={[]}
-      filters={filters}
-      pagination={{
-        endCursor: endCursor ?? undefined,
-        hasNextPage,
-        hasPreviousPage,
-        startCursor: startCursor ?? undefined,
-      }}
-      products={products}
-      sort={[
-        { value: 'featured', label: t('featuredItems'), selected: selectedSort === 'featured' },
-        { value: 'newest', label: t('newestItems'), selected: selectedSort === 'newest' },
-        {
-          value: 'best_selling',
-          label: t('bestSellingItems'),
-          selected: selectedSort === 'best_selling',
-        },
-        { value: 'a_to_z', label: t('aToZ'), selected: selectedSort === 'a_to_z' },
-        { value: 'z_to_a', label: t('zToA'), selected: selectedSort === 'z_to_a' },
-        {
-          value: 'best_reviewed',
-          label: t('byReview'),
-          selected: selectedSort === 'best_reviewed',
-        },
-        {
-          value: 'lowest_price',
-          label: t('priceAscending'),
-          selected: selectedSort === 'lowest_price',
-        },
-        {
-          value: 'highest_price',
-          label: t('priceDescending'),
-          selected: selectedSort === 'highest_price',
-        },
-        { value: 'relevance', label: t('relevance'), selected: selectedSort === 'relevance' },
-      ]}
-      title={searchTerm}
-      totalProducts={totalProducts}
-    />
+    <Suspense fallback={<CompareProductsSkeleton />}>
+      <ProductList searchParams={searchParams} searchTerm={searchTerm} />
+    </Suspense>
   );
 }
